Updates simplebully_skel-2.c
#20 Changes is_timeout() function to is_not_timeout(). Modifies it to
run MPI_Iprobe() to test if there was a run over of time. Fixes a
mistake I made, by forgetting to initialize the `msg[]` variable.
If there is no timeout, the message is taken out of the buffer, and the
type of msg is tested.
Right now, only the HELLO msg is filled in.